refactor: move `download.ts` file to `api` directory
